+++ packages/core/src/schema/input.ts
@@ -22,30 +22,44 @@ import {
   getGraphQLType,
   isSilk,
 } from "../resolver"
-import { mapValue, tryIn } from "../utils"
+import { mapValue, pascalCase, tryIn } from "../utils"
+import { LoomObjectType } from "./object"
 import type { CoreSchemaWeaverConfig } from "./types"
 import { provideWeaverContext, weaverContext } from "./weaver-context"
 
+interface EnsureInputOptions {
+  fieldName?: string
+}
+
 export function inputToArgs(
-  input: InputSchema<GraphQLSilk>
+  input: InputSchema<GraphQLSilk>,
+  options: EnsureInputOptions | undefined
 ): GraphQLFieldConfigArgumentMap | undefined {
   if (input === undefined) return undefined
   if (isSilk(input)) {
     let inputType = getGraphQLType(input)
     if (isNonNullType(inputType)) inputType = inputType.ofType
     if (isObjectType(inputType)) {
-      return mapValue(inputType.toConfig().fields, (it) =>
-        toInputFieldConfig(it)
-      )
+      return mapValue(inputType.toConfig().fields, (it, key) => {
+        let fieldName
+        if (options?.fieldName) {
+          fieldName = `${pascalCase(options.fieldName)}${pascalCase(key)}`
+        }
+        return toInputFieldConfig(it, { fieldName })
+      })
     }
     throw new Error(`Cannot convert ${inputType.toString()} to input type`)
   }
   const args: GraphQLFieldConfigArgumentMap = {}
   Object.entries(input).forEach(([name, field]) => {
     tryIn(() => {
+      let fieldName
+      if (options?.fieldName) {
+        fieldName = `${pascalCase(options.fieldName)}${pascalCase(name)}`
+      }
       args[name] = {
         ...field,
-        type: ensureInputType(field),
+        type: ensureInputType(field, { fieldName }),
       }
     }, name)
   })
@@ -53,7 +67,8 @@ export function inputToArgs(
 }
 
 export function ensureInputType(
-  silkOrType: GraphQLType | GraphQLSilk
+  silkOrType: GraphQLType | GraphQLSilk,
+  options: EnsureInputOptions | undefined
 ): GraphQLInputType {
   const gqlType = (() => {
     if (isSilk(silkOrType)) {
@@ -65,40 +80,44 @@ export function ensureInputType(
   if (isUnionType(gqlType))
     throw new Error(`Cannot convert union type ${gqlType.name} to input type`)
   if (isNonNullType(gqlType)) {
-    return new GraphQLNonNull(ensureInputType(gqlType.ofType))
+    return new GraphQLNonNull(ensureInputType(gqlType.ofType, options))
   }
   if (isListType(gqlType)) {
-    return new GraphQLList(ensureInputType(gqlType.ofType))
+    return new GraphQLList(ensureInputType(gqlType.ofType, options))
   }
   if (isObjectType(gqlType) || isInterfaceType(gqlType))
-    return ensureInputObjectType(gqlType)
+    return ensureInputObjectType(gqlType, options)
   return gqlType
 }
 
 export function ensureInputObjectType(
-  object: GraphQLObjectType | GraphQLInterfaceType | GraphQLInputObjectType
+  object: GraphQLObjectType | GraphQLInterfaceType | GraphQLInputObjectType,
+  options: EnsureInputOptions | undefined
 ): GraphQLInputObjectType {
   if (isInputObjectType(object)) return object
 
   const existing = weaverContext.inputMap?.get(object)
   if (existing != null) return existing
 
-  const {
-    astNode: _,
-    extensionASTNodes: __,
-    fields,
-    ...config
-  } = object.toConfig()
-
+  const { astNode, extensionASTNodes, fields, ...config } = object.toConfig()
+  let name = object.name
+  if (name === LoomObjectType.AUTO_ALIASING) {
+    name = `${pascalCase(options?.fieldName ?? "")}Input`
+  }
   const getInputObjectName =
     weaverContext.getConfig<CoreSchemaWeaverConfig>("gqloom.core.schema")
-      ?.getInputObjectName ?? ((name) => name)
+      ?.getInputObjectName ?? ((n) => n)
 
+  name = getInputObjectName(name)
   const input = new GraphQLInputObjectType({
     ...config,
-    name: getInputObjectName(object.name),
+    name,
     fields: provideWeaverContext.inherit(() =>
-      mapValue(fields, (it) => toInputFieldConfig(it))
+      mapValue(fields, (it, key) =>
+        toInputFieldConfig(it, {
+          fieldName: inputFieldName(name) + pascalCase(key),
+        })
+      )
     ),
   })
 
@@ -106,10 +125,16 @@ export function ensureInputObjectType(
   return input
 }
 
-function toInputFieldConfig({
-  astNode: _,
-  resolve: _1,
-  ...config
-}: GraphQLFieldConfig<any, any>): GraphQLInputFieldConfig {
-  return { ...config, type: ensureInputType(config.type) }
+function toInputFieldConfig(
+  { astNode, resolve, ...config }: GraphQLFieldConfig<any, any>,
+  options: EnsureInputOptions | undefined
+): GraphQLInputFieldConfig {
+  return { ...config, type: ensureInputType(config.type, options) }
+}
+
+function inputFieldName(name: string): string {
+  while (name.endsWith("Input")) {
+    name = name.slice(0, -"Input".length)
+  }
+  return name
 }