+++ packages/core/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+* Feature: auto assign names to objects and inputs
+
 ## 0.5.0 (2024-12-03)
 
 * Fix: add all objects in `resolver.of` to `context.types`